Document non-ideal behavior
If the file part of the path is a dangling symbolic link, one would, ideally, like for the symbolic link's destination to be created.
Instead, our implementation treats the symbolic link itself as a non-existing file, and replaces it with the new version. While not
ideal, this is a reasonable compromise, as the alternative is to rewrite realpath(3) to not fail on dangling end link.